Added the ability to select the built-in :lexical or :decimal comparison functions for B-tree databases, or even set a custom Proc to order the keys. This feature is currently only available through the Edo interface.
I did try to add it to the Tokyo side as well, but it's not possible without a
lot of work.  The reason is that comparison function must be set before the
B-tree database is opened.  Because the Tokyo interface uses the abstract
interface though, the underlying B-tree database isn't created until the open()
call.  Thus, the abstract interface would need to be abandoned for B-tree
databases to support this feature.